Meer omgezet naar Kotlin Foutje in grafiek per uur van de dag ivm zomertijd opgelost

+++ whdon_app/src/main/java/nl/whitedove/washetdroogofniet/Statistiek1Dag.kt
@@ -7,6 +7,6 @@ internal class Statistiek1Dag {
     var datum: DateTime? = null
     var aantalDroog: Int = 0
     var aantalNat: Int = 0
-    var minTemperatuur: Float = 0.toFloat()
-    var maxTemperatuur: Float = 0.toFloat()
+    var minTemperatuur: Int = 0
+    var maxTemperatuur: Int = 0
 }
